requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
matplotlib>=3.3.4
tensorboard>=2.7.0
PyYAML>=5.4.1
tqdm>=4.62.0